Acceptance tests and unit tests not yet

+++ src/main/java/com/acme/offirent/service/ReservationServiceImpl.java
@@ -4,6 +4,7 @@ import com.acme.offirent.domain.model.Reservation;
 import com.acme.offirent.domain.repository.AccountRepository;
 import com.acme.offirent.domain.repository.ReservationRepository;
 import com.acme.offirent.domain.service.ReservationService;
+import com.acme.offirent.exception.ResourceNotFoundException;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.data.domain.Page;
 import org.springframework.data.domain.PageImpl;
@@ -20,33 +21,32 @@ public class ReservationServiceImpl implements ReservationService {
     @Autowired
     private AccountRepository accountRepository;
 
-    @Override
-    public Page<Reservation> getAllReservations(Pageable pageable) {
-        return reservationRepository.findAll(pageable);
-    }
 
     @Override
-    public Reservation getReservationById(Long reservationId) {
-        return reservationRepository.findById(reservationId)
-                .orElseThrow(()->new ResourceNotFoundException("Reservation","Id",reservationId));
+    public Reservation getReservationByIdAndAccountId(Long accountId, Long reservationId) {
+        return reservationRepository.findByIdAndAccountId(reservationId, accountId)
+                .orElseThrow(() -> new ResourceNotFoundException(
+                        "Reservation not found with Id " + reservationId +
+                                " and AccountId " + accountId));
     }
 
     @Override
     public Page<Reservation> getAllReservationsByAccountId(Long accountId, Pageable pageable) {
-        return accountRepository.findById(accountId).map(account -> {
-            List<Reservation> reservations = account.getReservations();
-            int reservationsCount = reservations.size();
-            return new PageImpl<>(reservations,pageable,reservationsCount);
+      return reservationRepository.findByAccountId(accountId,pageable);
+    }
+
+    @Override
+    public Reservation createReservation(Long accountId, Reservation reservation) {
+        return accountRepository.findById(accountId).map(account->{
+            reservation.setAccount(account);
+            return reservationRepository.save(reservation);
         }).orElseThrow(()->new ResourceNotFoundException("Account","Id",accountId));
     }
 
     @Override
-    public Reservation createReservation(Reservation reservation) {
-        return reservationRepository.save(reservation);
-    }
-
-    @Override
-    public Reservation updateReservation(Long reservationId, Reservation reservationRequest) {
+    public Reservation updateReservation(Long accountId, Long reservationId, Reservation reservationRequest) {
+        if(!accountRepository.existsById(accountId))
+            throw new ResourceNotFoundException("Account","Id", accountId);
         return reservationRepository.findById(reservationId).map(reservation -> {
             reservation.setStatus(reservationRequest.getStatus());
             reservation.setInitialDate(reservationRequest.getInitialDate());
@@ -56,7 +56,10 @@ public class ReservationServiceImpl implements ReservationService {
     }
 
     @Override
-    public ResponseEntity<?> deleteReservation(Long reservationId) {
+    public ResponseEntity<?> deleteReservation(Long accountId, Long reservationId) {
+        if(!accountRepository.existsById(accountId))
+            throw new ResourceNotFoundException("Account","Id", accountId);
+
         return reservationRepository.findById(reservationId).map(reservation -> {
             reservationRepository.delete(reservation);
             return ResponseEntity.ok().build();
